AsyncBOT/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
discord-py==2.7.0a5350+g8f2cb607
frozenlist==1.5.0
idna==3.11
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.4
typing-extensions==4.13.2
urllib3==2.2.3
yarl==1.15.2
